filter people by academic appointment
